refactor: Update Meeting service and mapper to work with Meeting helper methods
The MeetingService was simplified by delegating the responsibility of linking child entities (Attendance, Strike, Participation point) to a new helper method within the Meeting entity itself.
The MeetingMapper was updated to use `CollectionMappingStrategy.ADDER_PREFERRED` to align with the new adder-based logic provided by the Meeting helper methods.
This change is more aligned with DDD prerrogatives.